Refactoring: two more newtypes, for AttribLocation and UniformLocation.
It would be nice to figure a HOAS-y way to make these safe, e.g. to
get some sort of guarantee that the identifiers I am talking about
actually correspond to something in the Shader program that we are
manipulating.  But for now this will do.

Anyway, the code is a little bit cleaner now.