Experiment: Reflect types to cleanup `vertex_attrib_pointer` API.
Some of this would be better with UFCS (and maybe associated items).

It also assumes that the vertices array is using a uniform type (which
seems plausible if you are not doing anything too clever).  If you
cannot provide that, then you have to use one of the other impls of
`ToVertexAttribPointerArgs`.

Also, even at its apex, its probably not as good as trying to put in a
higher-level encoding of what the layout of the vertices array
actually is.  For example, the stride is probably the same for all
entries in a given vertices array (right?  seems at least *likely* to
me...)

But I at least think the callers in this code are much easier to read.