Refactoring: made higher level interface to TexImage2D.
I attempted to guess what the protocol is here and how to convert from
SDL's pixel-format tag scheme to OpenGL's pixel-format scheme (which
requires in some cases for one to specify that the bytes need to be
reversed as they are read in ... I am somewhat amazed and wonder what
this expresses that could not have been handled by "just" doubling the
tag space instead.  Presumably there are other way more interesting
transformations one can perform on the data as it is being read in).