Attempted port of look_at and perspective and their application.
Note something is wrong with how I did look_at, but I cannot figure
out what it is.  For now I am just going to commit this work and try
to do a review of the matrix fundementals and then more direct
experimentation on what effects some adjustments have here.