Add `glsl::Packed` variant of `VertexAttribPointerRTTI`.
To further ease calls to `vertex_attrib_pointer`, make
`AttribLocation` carry the type of the value in the location, and then
enable one to indicate that the data-source that it is drawing from is
just a packed-array of such data.

Supporting this requires a little bit of acrobatics to allow the
`VertexAttribPointerRTTI` trait to actually make use of that location
type information when necessary (that's what the new type parameter on
`VertexAttribPointerRTTI` is for).

If I were really smart (and/or if Rust's type system were a bit more
expressive) I would also make the implementation of the `(&'a TUPLE,
&'a ROW_TYPE)` enforce that you only use compatible types such that
`TUPLE` can be assigned to `T`.  But that's okay, I can live without
that.