Demo usage of `layout (location = <n>)` qualifier.
This is a way to encode the attrib location directly in glsl rather
than by calling `glBindAttribLocation`.  (It is currently more verbose
in my code but I might put in ways to work around that, such as a new
variant of `fn global` that returns both a `Global` and an
`AttribLocation`.